Rework configuration system
