+++ packages/guider/package.json
@@ -92,6 +92,7 @@
     "approximate-number": "^2.1.1",
     "classnames": "^2.5.1",
     "color": "^4.2.3",
+    "defu": "^6.1.4",
     "extra-watch-webpack-plugin": "^1.0.3",
     "glob": "^10.3.10",
     "gray-matter": "^4.0.3",
