+++ packages/guider/src/client/index.ts
@@ -1,5 +1,4 @@
 export * from './virtuals';
-export * from './types';
 
 export * from './page/create-mdx-page';
 export * from './page/create-404-page';
