--- packages/guider/src/client/theme/components/component.ts
@@ -1,14 +1,0 @@
-import type { CustomComponentBuilder } from './types';
-
-export const component: CustomComponentBuilder = (componentOrOptions) => {
-  if (typeof componentOrOptions !== 'function') {
-    return {
-      ...componentOrOptions,
-      type: 'component',
-    };
-  }
-  return {
-    component: componentOrOptions,
-    type: 'component',
-  };
-};
