--- packages/guider/src/client/theme/components/directory.ts
@@ -1,9 +1,0 @@
-import type { DirectoryBuilder } from './types';
-
-export const directory: DirectoryBuilder = function (ops) {
-  return {
-    layoutSettings: {},
-    layout: 'default',
-    ...ops,
-  };
-};
