--- packages/guider/src/client/theme/components/seperator.ts
@@ -1,7 +1,0 @@
-import type { SeperatorBuilder } from './types';
-
-export const seperator: SeperatorBuilder = () => {
-  return {
-    type: 'seperator',
-  };
-};
