+++ packages/guider/src/client/virtuals.ts
@@ -2,7 +2,7 @@ import {
   sites as untypedSites,
   pageMap as untypedPageMap,
 } from '@neato/guider/shim.guider.virtual';
-import type { PageMapItem, PopulatedSiteConf } from './types';
+import type { PageMapItem, SiteComponent } from '../theme';
 
-export const sites = untypedSites as PopulatedSiteConf[];
+export const sites = untypedSites as SiteComponent[];
 export const pageMap = untypedPageMap as PageMapItem[];
