+++ packages/guider/src/theme.ts
@@ -1,3 +1,3 @@
-export * from './client/theme/index';
-export * from './client/theme/components';
-export * from './client/types';
+export * from './theme/index';
+export * from './theme/components/index';
+export * from './theme/types';
