+++ packages/guider/src/theme/components/index.ts
@@ -3,4 +3,7 @@ export * from './directory';
 export * from './group';
 export * from './link';
 export * from './seperator';
-export type * from './types';
+export * from './site';
+export * from './footer';
+export * from './layout';
+export * from './settings';
