+++ packages/guider/src/theme/components/seperator.ts
@@ -1,0 +1,11 @@
+export interface SeperatorComponent {
+  type: 'seperator';
+}
+
+export type SeperatorBuilder = () => SeperatorComponent;
+
+export const seperator: SeperatorBuilder = () => {
+  return {
+    type: 'seperator',
+  };
+};
