+++ packages/guider/src/theme/types.ts
@@ -1,0 +1,15 @@
+import type { SiteComponent } from './components/site';
+
+export type GuiderConfig = SiteComponent | SiteComponent[];
+
+export type MetaConf = {
+  site?: string;
+  directory?: string;
+  layout?: string;
+};
+
+export type PageMapItem = {
+  sitePath: string;
+  fileContents: Record<string, any>;
+  config: MetaConf;
+};
